Move TestContext and OwnedChild into modules
